requirements.txt
asgiref==3.11.0
django==6.0
sqlparse==0.5.5
python-dotenv
psycopg[binary]
google-genai
Pillow
pillow-heif
pydantic